+++ src/main/kotlin/dev/dediamondpro/resourcify/services/curseforge/CurseForgeMinecraftVersions.kt
@@ -20,8 +20,7 @@ package dev.dediamondpro.resourcify.services.curseforge
 import com.google.gson.annotations.SerializedName
 
 data class CurseForgeMinecraftVersions(
-    @SerializedName("versionString") val name: String,
-    @SerializedName("gameVersionTypeId") val id: Int
+    @SerializedName("versionString") val name: String
 )
 
 data class CurseForgeMinecraftVersionsResponse(val data: List<CurseForgeMinecraftVersions>)